move SearchDrawer to folder, create SearchField component
